Prevent duplicate channels, fix bugs
Hiding a channel from the recommended area also hid the channel avatar in their profile page and was fixed by adding a more detailed CSS selector. If the entire row in a channel was part of the filtered array of channels, then the row was collapsed. This was fixed by adding a height CSS property.